+++ src/commands/helpInfo.js
