+++ src/constants/smileys.js
